Added a parameter to details
The optional parameter of string called returnController, which is set to defualt "Game". This information will be passed to from a details from another controller, and then pass this as TempData to the game details view. This allows the back button to return to the details of the return controller passed.